DRYed up the orderfilter tests
